Redesigned upload parameters, added "--trigger-deadline" and "--save-blend"

+++ modules/render_output.py
@@ -35,6 +35,7 @@ def render(data: dict):
     output_file_video = output_dir / f"{render_name}.mp4"
     output_file_image = output_dir / f"{render_name}.png"
     output_file_audio = output_dir / f"{render_name}.wav"
+    output_file_blend = output_dir / f"{render_name}.blend"
     output_folder_jpg_sequence = output_dir / "jpg"
     output_folder_exr_sequence = output_dir / "exr" / f"Image"
 
@@ -113,29 +114,9 @@ def render(data: dict):
     else:
         print(f"Skipped rendering..")
 
-    # If the output folder exists, add the json file to it
-    if output_dir.exists():
-        shutil.copy(args_handler.json_path, output_dir)
-
-    # Upload the rendered result to S3 if any were generated
-    if args_handler.upload_render:
-        # Create a small file to indicate when the render is complete, it gets uploaded last
-        (output_dir / ".render_complete.txt").touch()
-
-        utils.upload_to_s3(output_dir, output_dir.parent)
-
-        # After the upload, delete the rendered output folder
-        if not args_handler.keep_files:
-            shutil.rmtree(output_dir, ignore_errors=True)
-
-    # Upload the blend file to S3
-    if args_handler.upload_blend:
-        # Create the output dir if it doesn't exist
-        output_file_blend = output_dir / f"{render_name}.blend"
+    if args_handler.save_blend:
+        # Create the output dir if it doesn't exist, Blender won't create it automatically when saving blend files
         output_dir.mkdir(parents=True, exist_ok=True)
-
-        # Create a small file to indicate when the render is complete, it gets uploaded last
-        (output_dir / ".render_complete.txt").touch()
 
         # Purge unused assets, pack all assets and save the blend file
         bpy.ops.outliner.orphans_purge(do_local_ids=True, do_linked_ids=True, do_recursive=True)
@@ -143,11 +124,28 @@ def render(data: dict):
         bpy.ops.wm.save_as_mainfile(filepath=str(output_file_blend))
         print(f"Saved blend file to: {output_file_blend}")
 
+    # If the output folder exists, add the json file to it
+    if output_dir.exists():
+        shutil.copy(args_handler.json_path, output_dir)
+
+    if args_handler.upload:
+        # Create a small file to indicate when the render is complete, it gets uploaded last
+        (output_dir / ".render_complete.txt").touch()
+
         utils.upload_to_s3(output_dir, output_dir.parent)
 
-        # After the upload, delete the rendered output folder
+    if args_handler.trigger_deadline:
+        if not output_file_blend.exists():
+            print(f"ERROR: No blend file for upload to Deadline found. Use '--save-blend' to save the blend file.")
+            return
+
+        utils.upload_to_s3(output_file_blend, output_dir, bucket_name="metabull-deadline-blend-files")
+
+    # After the upload, delete the output folder
+    if args_handler.upload or args_handler.trigger_deadline:
         if not args_handler.keep_files:
             shutil.rmtree(output_dir, ignore_errors=True)
+
